fileservice added!!!

+++ AniPart/InstruementMS/src/main/java/com/example/InstruementMS/instruFile/InstrumentController.java
@@ -1,6 +1,9 @@
 package com.example.InstruementMS.instruFile;
 
+import com.example.InstruementMS.external.User;
+import com.example.InstruementMS.external.UserClient;
 import com.example.InstruementMS.util.ApiResponse;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -11,18 +14,28 @@ import java.util.List;
 public class InstrumentController {
 
     @Autowired
-    InstrumentRepo instrumentRepo;
+    private InstrumentRepo instrumentRepo;
+    @Autowired
+    private UserClient userClient;
 
     @PostMapping
-    public ApiResponse<Instrument>createInstrument(@RequestBody Instrument instrument)
+    public ApiResponse<Instrument>createInstrument(@RequestBody Instrument instrument,
+                                                   HttpServletRequest request)
     {
+        User user = validateRequest(request);
+        instrument.setLastModifiedBy(user.getId());
         Instrument saved = instrumentRepo.save(instrument);
         return ApiResponse.onSuccess("Successfully created instrument", saved);
     }
 
     // Read (Get by ID)
     @GetMapping("/{id}")
-    public ApiResponse<Instrument> getInstrumentById(@PathVariable Long id) {
+    public ApiResponse<Instrument> getInstrumentById(@PathVariable Long id,
+                                                     HttpServletRequest request) {
+        User user = validateRequest(request);
+        if(user == null) {
+            return ApiResponse.onError("User not found");
+        }
         Instrument instrument = instrumentRepo.findById(id)
                 .orElseThrow(() -> new RuntimeException("Instrument not found"));
         return ApiResponse.onSuccess("Instrument fetched successfully", instrument);
@@ -30,14 +43,24 @@ public class InstrumentController {
 
     // Read All
     @GetMapping
-    public ApiResponse<List<Instrument>> getAllInstruments() {
+    public ApiResponse<List<Instrument>> getAllInstruments(HttpServletRequest request) {
+        User user = validateRequest(request);
+        if(user == null) {
+            return ApiResponse.onError("User not found");
+        }
         List<Instrument> instruments = instrumentRepo.findAll();
         return ApiResponse.onSuccess("All instruments fetched successfully", instruments);
     }
 
     // Update
     @PutMapping("/{id}")
-    public ApiResponse<Instrument> updateInstrument(@PathVariable Long id, @RequestBody Instrument updatedInstrument) {
+    public ApiResponse<Instrument> updateInstrument(@PathVariable Long id,
+                                                    @RequestBody Instrument updatedInstrument,
+                                                    HttpServletRequest request) {
+        User user = validateRequest(request);
+        if(user == null) {
+            return ApiResponse.onError("User not found");
+        }
         Instrument existing = instrumentRepo.findById(id)
                 .orElseThrow(() -> new RuntimeException("Instrument not found"));
 
@@ -45,6 +68,7 @@ public class InstrumentController {
         existing.setDescription(updatedInstrument.getDescription());
         existing.setActive(updatedInstrument.isActive());
         existing.setImage(updatedInstrument.getImage());
+        existing.setLastModifiedBy(user.getId());
         Instrument saved = instrumentRepo.save(existing);
         return ApiResponse.onSuccess("Instrument updated successfully",
                 saved);
@@ -52,11 +76,23 @@ public class InstrumentController {
 
     // Delete
     @DeleteMapping("/{id}")
-    public ApiResponse<String> deleteInstrument(@PathVariable Long id) {
+    public ApiResponse<Instrument> deleteInstrument(@PathVariable Long id,
+                                                    HttpServletRequest request) {
+        User user = validateRequest(request);
+        if(user == null) {
+            return ApiResponse.onError("User not found");
+        }
         Instrument existing = instrumentRepo.findById(id)
                 .orElseThrow(() -> new RuntimeException("Instrument not found"));
-
+        existing.setLastModifiedBy(user.getId());
         instrumentRepo.delete(existing);
-        return ApiResponse.onSuccess("Instrument deleted successfully", null);
+        return ApiResponse.onSuccess("Instrument deleted successfully", existing);
+    }
+
+    User validateRequest(HttpServletRequest request)
+    {
+        String bearerToken = request.getHeader("Authorization");
+        User user = userClient.verifyJWT(bearerToken).getData();
+        return user;
     }
 }
